更新版本

+++ vsound-provider/vsound-service/src/main/java/com/corn/vsound/service/code/strategy/code/CodeDeleteStrategy.java
@@ -3,18 +3,18 @@ package com.corn.vsound.service.code.strategy.code;
 import com.corn.boot.base.strategy.CudExecuteInterface;
 import com.corn.boot.error.BizError;
 import com.corn.vsound.dao.entity.CodeBase;
-import com.corn.vsound.dao.mapper.CodeBaseMapper;
-import com.corn.vsound.dao.mapper.CodeMethodMapper;
-import com.corn.vsound.dao.mapper.CodeOutSideUrlMapper;
-import com.corn.vsound.dao.mapper.CodeParameterMapper;
+import com.corn.vsound.dao.entity.CodeMethod;
+import com.corn.vsound.dao.mapper.*;
 import com.corn.vsound.facade.code.order.CodeCUDOrder;
 import com.corn.vsound.facade.code.result.CodeCUDResult;
+import com.corn.vsound.service.code.strategy.codemethod.CodeMethodDeleteStrategy;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.util.ObjectUtils;
 
 import java.util.Arrays;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class CodeDeleteStrategy implements CudExecuteInterface<CodeCUDOrder, CodeCUDResult> {
@@ -31,6 +31,9 @@ public class CodeDeleteStrategy implements CudExecuteInterface<CodeCUDOrder, Cod
     @Autowired
     private CodeOutSideUrlMapper codeOutSideUrlMapper;
 
+    @Autowired
+    private CodeMethodOrderMapper codeMethodOrderMapper;
+
     @Override
     public CodeCUDResult execute(CodeCUDOrder codeCUDOrder) {
 
@@ -41,13 +44,16 @@ public class CodeDeleteStrategy implements CudExecuteInterface<CodeCUDOrder, Cod
         }
 
         codeBaseMapper.deleteByPrimaryKey(codeId);
-
         List<String> codeIds = Arrays.asList(codeId);
-        codeMethodMapper.deleteCodeMethodsByCodeIds(codeIds);
+
+        List<CodeMethod> codeMethods = codeMethodMapper.findCodeMethodListByCodeId(codeIds);
+        if(!ObjectUtils.isEmpty(codeMethods)){
+            codeMethodMapper.deleteCodeMethodsByCodeIds(codeIds);
+            codeMethodOrderMapper.batchDeleteMethodOrder(codeMethods.stream().map(CodeMethod::getMethodId).collect(Collectors.toList()));
+        }
+
         codeParameterMapper.deleteCodeParametersByCodeIds(codeIds);
         codeOutSideUrlMapper.deleteCodeOutSideUrlsByCodeIds(codeIds);
         return new CodeCUDResult();
     }
-
-
 }
